requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
agentkit>=0.1.0
web3>=6.11.1
ccxt>=4.1.13
pandas>=2.1.4
numpy>=1.26.2
fastapi>=0.105.0
uvicorn>=0.24.0
python-jose>=3.3.0
websockets>=12.0
aiohttp>=3.9.1
pydantic>=2.5.2
flow-py-sdk>=1.1.0